+++ cmd/influx/main.go
@@ -90,6 +90,9 @@ type genericCLIOpts struct {
 	w    io.Writer
 	errW io.Writer
 
+	json        bool
+	hideHeaders bool
+
 	runEWrapFn cobraRunEMiddleware
 }
 
@@ -123,7 +126,13 @@ func (o genericCLIOpts) writeJSON(v interface{}) error {
 }
 
 func (o genericCLIOpts) newTabWriter() *internal.TabWriter {
-	return internal.NewTabWriter(o.w)
+	w := internal.NewTabWriter(o.w)
+	w.HideHeaders(o.hideHeaders)
+	return w
+}
+
+func (o *genericCLIOpts) registerPrintOptions(cmd *cobra.Command) {
+	registerPrintOptions(cmd, &o.hideHeaders, &o.json)
 }
 
 func in(r io.Reader) genericCLIOptFn {
@@ -289,6 +298,7 @@ func influxCmd(opts ...genericCLIOptFn) *cobra.Command {
 		cmdSetup,
 		cmdStack,
 		cmdTask,
+		cmdTelegraf,
 		cmdTemplate,
 		cmdApply,
 		cmdTranspile,
@@ -330,7 +340,7 @@ func defaultConfigPath() (string, string, error) {
 	if err != nil {
 		return "", "", err
 	}
-	return filepath.Join(dir, http.DefaultConfigsFile), dir, nil
+	return filepath.Join(dir, fs.DefaultConfigsFile), dir, nil
 }
 
 func getConfigFromDefaultPath() config.Config {
@@ -355,18 +365,18 @@ func migrateOldCredential() {
 		return // no need for migration
 	}
 
-	tokB, err := ioutil.ReadFile(filepath.Join(dir, http.DefaultTokenFile))
+	tokB, err := ioutil.ReadFile(filepath.Join(dir, fs.DefaultTokenFile))
 	if err != nil {
 		return // no need for migration
 	}
 
-	err = writeConfigToPath(strings.TrimSpace(string(tokB)), "", filepath.Join(dir, http.DefaultConfigsFile), dir)
+	err = writeConfigToPath(strings.TrimSpace(string(tokB)), "", filepath.Join(dir, fs.DefaultConfigsFile), dir)
 	if err != nil {
 		return
 	}
 
 	// ignore the remove err
-	_ = os.Remove(filepath.Join(dir, http.DefaultTokenFile))
+	_ = os.Remove(filepath.Join(dir, fs.DefaultTokenFile))
 }
 
 func writeConfigToPath(tok, org, path, dir string) error {